+++ lib/opal/jquery/document.rb
@@ -1,5 +1,5 @@
-require 'opal-jquery/constants'
-require 'opal-jquery/element'
+require 'opal/jquery/constants'
+require 'opal/jquery/element'
 
 module Browser
   # {Document} includes these methods to extend {Element}.