+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 require 'opal-rspec'
-require 'opal-jquery'
-require 'opal-jquery/rspec'
+require 'opal/jquery'
+require 'opal/jquery/rspec'
 
 module JQueryTestHelpers
   def find(selector)